Fix the test case!
I have verified that with libxml2, the only reason you get those
"\n" characters before/after the comment node is that for testSAX.c,
there is not nodePush() call, which sets ctxt->node

This is the only clause that's makeing areBlanks() return false
(by ctx->node == NULL). This doesn't happen if the calls to nodePush()
happens as in helium.

The difference is caused by the fact that libxml2's nodePush requires
are real node (which is only generated in the SAX handler), whereas
helium uses a pseudo node (for now), so you can always push the node
on the stack, even without real SAX handler (i.e. TreeBuilder)